requirement.txt
streamlit
langchain_core
langchain_groq
langchain_community
Wikipedia-API
duckduckgo-search
youtube-transcript-api
sentence-transformers
chromadb
google
reportlab 
markdown 
beautifulsoup4
langchain
langchain-groq
pypdf
graphviz
python-dotenv
